Move getCodeOfConductTemplate
