+++ feature/Messages/Sources/Messages/Services/MessagesService/MessagesServiceImpl.swift
@@ -97,6 +97,26 @@ class MessagesServiceImpl: MessagesService {
         }
     }
 
+    struct GetCodeOfConductTemplateRequest: APIRequest {
+        typealias Response = RawResponse
+
+        var method: HTTPMethod { .get }
+
+        var resourceName: String {
+            "api/files/templates/code-of-conduct"
+        }
+    }
+
+    func getCodeOfConductTemplate() async -> DataState<String> {
+        let result = await client.sendRequest(GetCodeOfConductTemplateRequest())
+        switch result {
+        case .success((let rawResponse, _)):
+            return .done(response: rawResponse.rawData)
+        case .failure(let error):
+            return .failure(error: .init(error: error))
+        }
+    }
+
     struct GetCodeOfConductAgreementRequest: APIRequest {
         typealias Response = Bool
