+++ feature/Messages/Sources/Messages/Services/_Service.swift
@@ -13,7 +13,6 @@ import UserStore
 
 // swiftlint:disable:next type_name
 protocol _Service {
-    func getCodeOfConductTemplate() async -> DataState<String>
     func getCodeOfConductAgreement(for course: Course) async -> DataState<Bool>
     func acceptCodeOfConduct(for course: Course) async -> NetworkResponse
 }
@@ -25,27 +24,6 @@ enum _ServiceFactory {
 
 // swiftlint:disable:next type_name
 struct _ServiceImpl: _Service {
-    // swiftlint:disable:next type_name
-    struct _Request: APIRequest {
-        typealias Response = RawResponse
-
-        var method: HTTPMethod { .get }
-
-        var resourceName: String {
-            "api/files/templates/code-of-conduct"
-        }
-    }
-
-    func getCodeOfConductTemplate() async -> DataState<String> {
-        let result = await APIClient().sendRequest(_Request())
-        switch result {
-        case .success((let rawResponse, _)):
-            return .done(response: rawResponse.rawData)
-        case .failure(let error):
-            return .failure(error: .init(error: error))
-        }
-    }
-
     func getCodeOfConductAgreement(for course: Course) async -> DataState<Bool> {
         guard let serverHost = UserSession.shared.institution?.baseURL?.absoluteString else {
             return .failure(error: .init(title: "No base URL"))
